+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- New Wellness home panel (Task in progress) [#1689](https://github.com/rokwire/illinois-app/issues/1689).
 
 ## [4.0.22] - 2022-06-03
 ### Changed